updated package, readme, index
