+++ README.md
@@ -4,6 +4,7 @@ Easy-to-use PostgreSQL CRUD Handlers + More (Node)
 
 ## TODO
 
-- [ ] Show Examples of easy to create Table CRUD handlers
+- [ ] Provide Example Project to show CRUDBuilder usage on PostgreSQL models, use of builder utility functions and helpers
 - [ ] How It Works
-- [ ] Why Should You Use This?
+- [ ] Add Tests
+- [ ] "Why Use This?" Section
